requirements.txt
paho-mqtt
chromadb
ollama
streamlit
plotly
pandas
httpx